D-1-fix: typecheck projects without behaviour or dependency drift
Restore conditional Ponder mode/gateway registration, split Node tooling
from DOM test projects, and drop the Solana override/spl-token pin while
keeping every TypeScript file in exactly one gated project.

+++ test/bonded-challenge/BondedChallenge.test.ts
@@ -413,7 +413,7 @@ describe("BondedChallenge (CH1–CH6)", () => {
       revertsWith("CannotRouteBondToJudge"),
     );
 
-    assert.ok((await misconfiguredSettlement.read.challengeOpenedAt([subject])) > 0n);
+    assert.ok((await misconfiguredSettlement.read.challengeOpenedAt([subject]) as bigint) > 0n);
     assert.equal(await misconfiguredSettlement.read.pendingClaims([rejectJudge.address, ZERO]), 0n);
   });
